refactor: layout body
